+++ elastiknn-plugin/src/main/scala/com/klibisz/elastiknn/codec/ElastiknnCodecService.scala
@@ -6,7 +6,7 @@ import org.elasticsearch.index.codec.CodecService
 class ElastiknnCodecService extends CodecService(null, null) {
 
   override def codec(name: String): Codec =
-    Codec.forName(ElastiknnCodecService.ELASTIKNN_87)
+    Codec.forName(ElastiknnCodecService.ELASTIKNN_88)
 
 }
 
@@ -14,4 +14,5 @@ object ElastiknnCodecService {
   val ELASTIKNN_84 = "Elastiknn84Codec"
   val ELASTIKNN_86 = "Elastiknn86Codec"
   val ELASTIKNN_87 = "Elastiknn87Codec"
+  val ELASTIKNN_88 = "Elastiknn88Codec"
 }