+++ elastiknn-plugin/src/main/scala/com/klibisz/elastiknn/query/KnnScoreFunctionBuilder.scala
@@ -7,7 +7,7 @@ import com.klibisz.elastiknn.api.{NearestNeighborsQuery, Vec}
 import org.elasticsearch.common.io.stream.{StreamInput, StreamOutput, Writeable}
 import org.elasticsearch.common.lucene.search.function.ScoreFunction
 import org.elasticsearch.common.xcontent.{ToXContent, XContentBuilder, XContentParser}
-import org.elasticsearch.index.query.QueryShardContext
+import org.elasticsearch.index.query.SearchExecutionContext
 import org.elasticsearch.index.query.functionscore.{ScoreFunctionBuilder, ScoreFunctionParser}
 
 final class KnnScoreFunctionBuilder(val query: NearestNeighborsQuery, val weight: Float)
@@ -26,7 +26,7 @@ final class KnnScoreFunctionBuilder(val query: NearestNeighborsQuery, val weight
 
   override def doHashCode(): Int = Objects.hash(query, weight.asInstanceOf[java.lang.Float])
 
-  override def doToFunction(context: QueryShardContext): ScoreFunction = {
+  override def doToFunction(context: SearchExecutionContext): ScoreFunction = {
     ElastiknnQuery(query, context).map(_.toScoreFunction(context.getIndexReader)).get
   }
 }