+++ elastiknn-testing/src/test/scala/com/klibisz/elastiknn/ClusterSpec.scala
@@ -32,7 +32,7 @@ class ClusterSpec extends AsyncFunSuite with Matchers with Elastic4sMatchers wit
     } yield {
       catNodesRes.shouldBeSuccess
       catNodesRes.result should have length 2
-      catNodesRes.result.map(_.nodeRole).sorted shouldBe List("cdhilrstw", "mr")
+      catNodesRes.result.map(_.nodeRole).sorted shouldBe List("cdfhilrstw", "mr")
     }
   }